requirements.txt
arabic-reshaper==2.1.1
astroid==2.4.2
cairocffi==1.2.0
CairoSVG==2.5.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
Click==7.0
colorama==0.4.3
cryptography==3.4.7
cssselect2==0.4.1
defusedxml==0.6.0
diff-match-patch==20200713
Django==2.2.6
django-admin-interface==0.14.0
django-colorfield==0.3.2
django-easy-pdf==0.1.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-import-export==2.5.0
django-jazzmin==2.4.7
et-xmlfile==1.1.0
future==0.18.2
gTTS==2.2.1
html5lib==1.1
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
MarkupPy==1.14
mccabe==0.6.1
mysql-connector==2.2.9
mysql-connector-python==8.0.21
mysqlclient==1.4.6
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.7
Pillow==7.1.2
protobuf==3.12.2
psycopg2==2.8.4
pycparser==2.20
PyJWT==2.1.0
pylint==2.5.3
PyPDF2==1.26.0
Pyphen==0.10.0
PyQt5==5.11.3
PyQt5-sip==4.19.18
pyqt5-tools==5.11.3.1.4
python-bidi==0.4.2
python-dotenv==0.10.3
python3-openid==3.2.0
pytz==2019.3
PyYAML==5.4.1
razorpay==1.2.0
reportlab==3.5.56
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.3.0
tablib==3.0.0
tinycss2==1.1.0
toml==0.10.1
twilio==6.59.0
typed-ast==1.4.1
urllib3==1.26.2
views==0.3
virtualenv==16.6.2
virtualenvwrapper-win==1.2.5
WeasyPrint==52.2
webencodings==0.5.1
wrapt==1.12.1
xhtml2pdf==0.2.5
xlrd==2.0.1
xlwt==1.3.0
